consumer/requirements.txt
fastapi
redis
rq
uvicorn
python-dotenv